Group_XX_Hybrid_RAG/requirements.txt
sentence-transformers
transformers
torch
numpy
pandas
faiss-cpu
rank_bm25
streamlit
matplotlib
seaborn
pyyaml
tqdm
beautifulsoup4
requests
notebook
scikit-learn
nltk
rouge-score
bert-score
wikipedia-api
